+++ packages/client/src/app/threejs/dummy-walker.ts
