+++ packages/client/src/app/threejs/fps-counter.ts
